+++ addon/inspector.js
@@ -1,4 +1,5 @@
-export let apiVersion = localStorage.getItem("apiVersion") == null ? "61.0" : localStorage.getItem("apiVersion");
+export let defaultApiVersion = "61.0";
+export let apiVersion = localStorage.getItem("apiVersion") == null ? defaultApiVersion : localStorage.getItem("apiVersion");
 export let sessionError;
 export function nullToEmptyString(value) {
   // For react input fields, the value may not be null or undefined, so this will clean the value